Refactors most of the current api endpoints into a few different blueprints, and registers them on the app.
Spot-checked the majority of endpoints so should be OK to use, need to deploy to server and put all the endpoints into the tool to make sure it all actually still works.
